Remove one question, combine two in Maternity Pay
Question 2 for "Does the employee have an employment contract with you?"
is redundant as that can be inferred from other questions. This is now
removed.

Questions 4 and 5 are combined and their replacement becomes the new
Question 3.

As mentioned in the previous commit, instead of asking separately:

Q4 "Did the employee work for you on or before 1 April 2017?"

and

Q5 "Was the employee (or will they be) on your payroll on 17 September
2017?"

We will instead ask:

Q3 "Did the employee have a contract with you between 1 April 2017 and
17 September 2017?"

As the original 2 questions of 4 and 5 had 2 separate negative outcomes:

* `not_worked_long_enough`
* `must_be_on_payroll`

- these are now combined
into a new `:not_worked_long_enough_and_not_on_payroll`

Answering "yes" to the new Question 3 will continue with the rest of the
flow which is unchanged.