Fix approximation error in Statutory Maternity Pay
An approximation error was highlighted on the Statutory Maternity Pay
calculator, regarding average weekly earnings.

The calculator was 22p out in the calculation for someone paid 4
weekly.

A person with an income of £1187.24 in the relevant period should be
entitled to earn £5209.02 and not £5209.24, as it was given by the
Gov.uk Calculator.

1187.24 / 8 = 148.405
148.405 * 90% = 133.5645
133.5645 * 39 = 5209.02 rounded up to the nearest penny.

Zendesk ticket: https://govuk.zendesk.com/agent/tickets/2443078